BQ_queries_ETL/requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.5
blinker==1.6.2
build==0.10.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.4
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-bigquery==3.11.4
google-cloud-bigquery-datatransfer==3.12.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.16.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter_core==5.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==23.2.1
pip-tools==7.2.0
pipreqs==0.4.13
platformdirs==3.10.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
rsa==4.9
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
wheel==0.38.4
yarg==0.1.9
zipp==3.16.2
